chore: move logprob to a separate module
