+++ aphrodite/common/sequence.py
@@ -16,6 +16,7 @@ import torch
 from aphrodite.common.pooling_params import PoolingParams
 from aphrodite.common.sampling_params import RequestOutputKind, SamplingParams
 from aphrodite.inputs import SingletonInputs
+from aphrodite.logprobs import Logprob, PromptLogprobs, SampleLogprobs
 from aphrodite.multimodal import MultiModalKwargs, MultiModalPlaceholderDict
 
 if TYPE_CHECKING:
@@ -34,30 +35,6 @@ APHRODITE_INVALID_TOKEN_ID = -1
 def array_full(token_id: int, count: int):
     """[`array`][] equivalent of [numpy.full][]."""
     return array(APHRODITE_TOKEN_ID_ARRAY_TYPE, [token_id]) * count
-
-
-# We use dataclass for now because it is used for
-# openai server output, and msgspec is not serializable.
-# TODO(sang): Fix it.
-@dataclass
-class Logprob:
-    """Infos for supporting OpenAI compatible logprobs and token ranks.
-
-    Attributes:
-        logprob: The logprob of chosen token
-        rank: The vocab rank of chosen token (>=1)
-        decoded_token: The decoded chosen token index
-    """
-    logprob: float
-    rank: Optional[int] = None
-    decoded_token: Optional[str] = None
-
-
-# {token_id -> logprob} per each sequence group. None if the corresponding
-# sequence group doesn't require prompt logprob.
-PromptLogprobs = list[Optional[dict[int, Logprob]]]
-# {token_id -> logprob} for each sequence group.
-SampleLogprobs = list[dict[int, Logprob]]
 
 
 class SequenceStatus(enum.IntEnum):
